# Conflicts:
#	src/main/java/com/company/web/smart_garage/controllers/UserController.java
#	src/main/java/com/company/web/smart_garage/data_transfer_objects/UserDtoOut.java
#	src/main/java/com/company/web/smart_garage/data_transfer_objects/VisitDtoOutSimple.java
#	src/main/java/com/company/web/smart_garage/services/impl/UserServiceImpl.java
#	src/main/java/com/company/web/smart_garage/utils/helpers/VisitMapper.java

+++ src/main/java/com/company/web/smart_garage/services/UserService.java
@@ -30,7 +30,7 @@ public interface UserService {
 
     void delete(long id, User user);
 
-    void makeAdmin(int id, User user);
+    void makeAdmin(long id, User user);
 
     void makeEmployee(int id, User user);
